refactor caliban implementation
