+++ implementations/caliban/src/main/scala/models/DeprecatedProductArgs.scala
@@ -1,0 +1,10 @@
+package models
+
+import caliban.schema.{ArgBuilder, Schema}
+
+case class DeprecatedProductArgs(sku: String, `package`: String)
+
+object DeprecatedProductArgs {
+  implicit val schema: Schema[Any, DeprecatedProductArgs]    = Schema.gen
+  implicit val argBuilder: ArgBuilder[DeprecatedProductArgs] = ArgBuilder.gen
+}
