+++ implementations/caliban/src/main/scala/models/Inventory.scala
@@ -1,0 +1,16 @@
+package models
+
+import caliban.schema.{ GenericSchema, Schema }
+import services.{ InventoryService, UserService }
+
+@GQLInterfaceObject
+@GQLKey("email")
+case class Inventory(
+  id: ID,
+  deprecatedProducts: List[DeprecatedProduct]
+)
+
+object Inventory {
+  object genSchema extends GenericSchema[InventoryService with UserService]
+  implicit val schema: Schema[InventoryService with UserService, Inventory] = genSchema.gen
+}
