+++ implementations/caliban/src/main/scala/models/ProductResearch.scala
@@ -1,0 +1,13 @@
+package models
+
+import caliban.schema.Schema
+
+@GQLKey("study { caseNumber }")
+case class ProductResearch(
+  study: CaseStudy,
+  outcome: Option[String]
+)
+
+object ProductResearch {
+  implicit val schema: Schema[Any, ProductResearch] = Schema.gen
+}
