+++ implementations/caliban/src/main/scala/models/ProductResearchArgs.scala
@@ -1,0 +1,10 @@
+package models
+
+import caliban.schema.{ ArgBuilder, Schema }
+
+case class ProductResearchArgs(study: CaseStudyArgs)
+
+object ProductResearchArgs {
+  implicit val schema: Schema[Any, ProductResearchArgs]    = Schema.gen
+  implicit val argBuilder: ArgBuilder[ProductResearchArgs] = ArgBuilder.gen
+}
