+++ implementations/caliban/src/main/scala/models/UserArgs.scala
@@ -1,0 +1,10 @@
+package models
+
+import caliban.schema.{ ArgBuilder, Schema }
+
+case class UserArgs(email: ID)
+
+object UserArgs {
+  implicit val schema: Schema[Any, UserArgs]    = Schema.gen
+  implicit val argBuilder: ArgBuilder[UserArgs] = ArgBuilder.gen
+}
