+++ implementations/caliban/src/main/scala/models/package.scala
@@ -1,0 +1,10 @@
+import caliban.federation.v2x._
+
+package object models
+    extends FederationV2(
+      Versions.v2_3 :: List(
+        Link("https://myspecs.dev/myCustomDirective/v1.0", List(Import("@custom"))),
+        ComposeDirective("@custom")
+      )
+    )
+    with FederationDirectivesV2_3
